agent/requirements.txt
# OPAS Agent Runtime — Python 3.11+
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1

# Data access
supabase>=2.15.0
pgvector==0.3.4

# LLM + embeddings
anthropic==0.34.2
openai==1.51.0
google-generativeai>=0.8.0

# Corpus ingestion
pypdf==5.0.1
tiktoken==0.8.0

# YAML handling (skill YAML is stored as JSONB but locally we may also load from disk)
PyYAML==6.0.2

# File upload (form parsing)
python-multipart>=0.0.9

# Dev / tests
pytest==8.3.3
httpx==0.27.2
